+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `EyecandyBlockWrapper.redstoneLevel()` now returns the true redstone level, instead of only 0 and 15.
 - **PIDS Scripting:**
   - Add `TextWrapper.renderType` to specify a custom render type for text, alongside TextureWrapper.
+  - Add `TextWrapper.naturalLight()` and `TextureWrapper.naturalLight()` to allow rendering PIDS elements at world light. 
 
 ## Changes
 - Incorrect invocation of `ModelManager.upload()` will now throw an error instead of crashing the game.